community/gtk4.0: move gstreamer backend to -media-gstreamer subpkg
This component depends on gst-plugins-bad which is a pretty huge
dependency. And it's needed only for some applications. For example,
Debian provides it in a separate package (libgtk-4-media-gstreamer) too.

gtk4.0-media-gstreamer should be added as a dependency to the packages
that actually need it. Unfortunately, I don't know which ones are they
at this moment.