community/py3-qt6: check python dependency info in prepare()
This commit introduces the _check_dep() function which is used in
prepare() to read python dependency information from the pyproject.toml
and PKG-INFO files, and check it against the versions of the
dependencies installed.

Yes its hacky, but I keep forgetting that riverbank doesn't put any
useful dependency version information in the release notes. They seem to
rely entirely on their python toolchain to get it right, which does not
work with our build process. With this code I will at least be pointed
in the right direction the next time the dependency chain breaks.