testing/mepo: don't depend on a specific font
mepo-scripts uses icons from nerd fonts in its user interface. For this
to work properly, it needs to depend on a font that provides these
symbols.

Currently, mepo-scripts depends on font-hack-nerd. When font-hack-nerd
is installed, it will affect which font fontconfig chooses as a fallback
in other applications.

Change the dependency to font-nerd-fonts-symbols, which ships only the
nerd font symbols. This covers the need in mepo-scripts without
affecting any other use case.