main/busybox: include fixes for sed regex delimiter parsing
Similar to the awk fix, which also isn't upstream yet and was
backported from the ML, this can cause subtle hard-to-debug
breakage in software buildsystems which make use of sed(1).
Hence, it's useful to include it even if isn't upstream yet.

See: https://lists.busybox.net/pipermail/busybox/2024-November/091023.html