community/texlive: enable on riscv64,loongarch64 without luajit
TeX Live's bundled luajit doesn't support these architectures and it's
non-trivial to use a system-provided luajit. Most functionality in the
package works without luajit, though. Even the lua-enabled parts of
texlive have a non-JIT version.

With a small patch, it possible to disable LuaJIT in texlive.

On s390x, texlive also builds but xelatex fails to compile the sample
document.