community/nim: upgrade to 2.2.4, link compiler sources, fix stdlib location and debug tool packaging
This patch upgrades to Nim 2.2.4 and makes three packaging changes:

1. Corrects the location of the nim standard library to $nimdir/lib as
   per https://github.com/nim-lang/Nim/pull/24460 (accepted upstream).

2. Relocates the tools in tools/debug to the -dbg package

3. Adds a symlink to the compiler sources at $nimdir/compiler. The
   Nimble package manager requires these sources to compile when built
   separately from the compiler, and other distributions package these
   with the compiler.

The overall intent is to allow easier packaging of other tools (in
particular the nimble package manager and the nim LSP language server)
but the opportunity is taken to improve alignment with the upstream
packaging guidance (https://nim-lang.org/docs/packaging.html) where
possible.

The nimUse64BitCTime conditional  is only used in one place to alter the
definition of the CTime type, so just unconditionally define it as an
int64 and remove the conditional from the build steps.

The location of nim libs has been corrected upstream, so remove this fix
from the relevant installer script patch.

Finally, nodejs and pcre now  appear to be needed during the build
process, so these are moved from checkdepends to makedepends (checks
are disabled on some platforms and this was causing the build to fail).

--
Changelog 2.2..2.4:
- Upstream C sources now using time_t for unix times so C code patch no longer required.