main/nginx: add the creation of a subpackage `nginx-src`
The goal if to allow building of any nginx module without
having to modify nginx' APKBUILD, be it to package them later
or for testing purpose.

A special global variable `_configure_common` is created with all
the relevant `configure` script parameters and saved along the
source in the specific `nginx-src` package. This permits to
compile a new module using the original package parameter
signature and be "binary compatible" without the need to use the
`--with-compat` trick.