WIP: testing/archivebox: new aport
This builds, but crashes soon after start-up. The root of the crash is
ArchiveBox depending on an older, incompatible version of Django.

The latest development tree seems to use a newer Django, as well as
several other dependencies absent in the aports tree.

It's probably best to continue this work once upstream releases 0.8.0,
unless someone wants to package track the current development branch.

Committing this WIP version so that the work isn't lost.