main/python3.12: support altinstall
This allows installing multiple python versions side-by-side on end-user
installations. The goal is to allow to setup a seperate Python
virtualenv with a different Python version than Alpine's default.

As an additional benefit, I hope that eases the effort to switch for
Alpine Linux to switch to a different default version.

Official Alpine packages should still build against the default python
version, and not depend on alternative python versions.

I don't intend to make the maintainance of it harder by any way with this,
and i hope this goes in the other direction and helps maintaining it.