main/fontconfig: don't update font-cache as root
When updating the font cache, run the operation as an unprivileged user
instead of the root user.

This is a second iteration of
https://gitlab.alpinelinux.org/alpine/aports/-/merge_requests/87404

Update permissions to the cache directory via pre-upgrade, rather than
post-upgrade , so that it is writeable before the trigger is in place.

Also update permissions for files inside the cache directory, which may
be present due to an existing or previous installation.